perf<Stack>:优化顺序栈的长度计算
- int64_* 强转 int，抑制精度丢失的警告

+++ Stack/SequenceStack.cpp
@@ -20,7 +20,7 @@ bool SequenceStack<T>::Empty()
 template<class T>
 int SequenceStack<T>::Length()
 {
-    return top - base;
+    return (int) (top - base);
 }
 
 template<class T>
